backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
idna==3.7
Markdown==3.5.2
mysqlclient==2.2.1
pillow==10.3.0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
